Add the `result` inside the `CompilationResult`
