+++ lib/tooling/base-tool.ts
@@ -31,12 +31,12 @@ import {ExecutionOptions} from '../../types/compilation/compilation.interfaces';
 import {UnprocessedExecResult} from '../../types/execution/execution.interfaces';
 import {Library, SelectedLibraryVersion} from '../../types/libraries/libraries.interfaces';
 import {ResultLine} from '../../types/resultline/resultline.interfaces';
+import {ToolInfo, ToolResult, ToolTypeKey} from '../../types/tool.interfaces';
 import * as exec from '../exec';
 import {logger} from '../logger';
 import {parseOutput} from '../utils';
 
-import {ITool, ToolEnv, ToolResult} from './base-tool.interface';
-import {ToolInfo, ToolTypeKey} from '../../types/tool.interfaces';
+import {ITool, ToolEnv} from './base-tool.interface';
 
 const toolCounter = new PromClient.Counter({
     name: 'tool_invocations_total',
